pyproject.toml
[project]
name = "azurecosmosdbfastapi"
version = "0.1.0"
description = "MicroService for Deploying AzureCosmosDB"
authors = [
    {name = "rijo.john@infracloud.io"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-mgmt-cosmosdb (>=9.7.0,<10.0.0)",
    "azure-identity (>=1.19.0,<2.0.0)",
    "fastapi (>=0.115.7,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "google-auth (>=2.38.0,<3.0.0)",
    "google-auth-oauthlib (>=1.2.1,<2.0.0)",
    "google-auth-httplib2 (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "pydantic-settings (>=2.7.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-mock = "^3.14.0"
httpx = "^0.28.1"
requests-mock = "^1.12.1"
pytest-asyncio = "^0.25.2"
mypy = "^1.14.1"
types-requests = "^2.32.0.20241016"
types-python-dateutil = "^2.9.0.20241206"


[tool.mypy]
strict = true
ignore_missing_imports = true
exclude = "^(venv|.venv|env|.env)"
warn_unused_configs = true
plugins = ["pydantic.mypy"]


[[tool.mypy.overrides]]
module = [
    "fastapi.*",
    "pydantic.*",
    "azure.*"
]
ignore_missing_imports = true
